More cleanup applied :
Delete the crc32file function from crc.c. This is a test function, that is not
exported and thus not used anywhere. It also deletes the main function there
that is commented out.

Remove the "/bin/sh" shebang in the sourced scripts in the ./scripts
directory. These files are not meant for direct execution and are installed
in /usr/lib/ anyway. This fix comes from the debian package dir (see next
patch).

Remove the now unnecessary patch in the debian packaging directories.

Adds some documentation for the replace callback in the api.txt documentation
in the db dir. It also extends the comment for this function in db.h.

Credits go to Henning Westerholt <henning.westerholt@1und1.de>


git-svn-id: https://opensips.svn.sourceforge.net/svnroot/opensips/trunk@1687 689a6050-402a-0410-94f2-e92a70836424