- bug fixed: Matching operators fail to match on some platforms. - Decription: Due to a bug in route.c, the regular expression matching operators do not work in all environments. For example, when building 1.4.0 or 1.3.2 on Solaris 10 (sun4v or x86) they give the same result independent of input. It would seem that if it were not for the accident of a particular struct alignment, these operators would be broken on all platforms.
The underlying problem is a bad test in comp_s2s, which expects s2 to be a
str*, and tests st->s==NULL. When handling the matching operators MATCH_OP
and NOTMATCH_OP, it is a regex_t* instead - a fact reflected in the cast
under the appropriate switch case.

Credits go to Amy Linari (patch a bit reworked). Closes patch 2044475.



git-svn-id: https://opensips.svn.sourceforge.net/svnroot/opensips/trunk@4599 689a6050-402a-0410-94f2-e92a70836424