- added two new helper functions for searching for non standard headers: 	get_header_by_name( msg, name, len)  	get_header_by_static_name( msg, static_name)
Second function is actually a macro over the first one - it is useful when the header name is a static string (and uses sizeof to get the len).

The new functions are to replace the local implementations form the modules and to fix the following bugs:
	1) header names are case insensitive 
	2) check also len (otherwise hdr ABCDE will match while looking for ABC)
	3) do string comp only if hdr type (OTHER) and name len match - more efficient

Partial fix for bug #2202134.



git-svn-id: https://opensips.svn.sourceforge.net/svnroot/opensips/trunk@4833 689a6050-402a-0410-94f2-e92a70836424