- reduce the number of memory fragements used by a DB result:    previous number = 1 + 2 + no_cols + 1 + no_rows    current number = 3
the number of mem fragments is now constant and what is more important is the fact that we avoid keep allocating very small chunks (like 8 bytes). 

This will reduce the fragmentation of the private memory due of continues DB queries -> opensips can run for longer periods of time with no mem issues.

Another commit will follow for syncronizing the DB implementation with these changes in the API



git-svn-id: https://opensips.svn.sourceforge.net/svnroot/opensips/trunk@5230 689a6050-402a-0410-94f2-e92a70836424