History behind this change (WARN -> INFO): The messages you are seeing means that the TCP MANAGER process sees that all the TCP WORKER processes are already processing other messages, so  no one is free (idle), so , it will queue the current active connection to one of the TCP WORKERs... There is nothing wrong with this, it is the normal way it works - of course, if you have too few TCP workers, the probability to have queuing is higher.
In UDP the queueing is not visible at application level as it is done by the kernel in the receive buffer that is assigned to the socket.

In TCP the queueing is visible as there is a single TCP MANAGER process managing the connections (detecting the reads, connects, etc) which is quite fast (not doing any processing) -> the queueing is moved between the TCP MANAGER and TCP WORKERS.

Also, to be sure you got it right, there is no relation between the TCP WORKER and a connection. A TCP WORKER process is just reading and processing a SIP message at a certain time. Next SIP message from the same connection may end up in a different TCP WORKER proc.



git-svn-id: https://opensips.svn.sourceforge.net/svnroot/opensips/trunk@6561 689a6050-402a-0410-94f2-e92a70836424