When rewriting the SDP of an ACK rtpproxy_answer() retrieves the wrong port from rtpproxy.
This is reproducible by handling SDP offered in a 200 Ok response with rtpproxy_offer() and then processing the answering SDP in an ACK with rtpproxy_answer() - the result is that the same port gets written into the ACK's SDP as was written into the Ok's SDP. In short, the 200 Ok is being handled correctly when it contain an SDP offer since the to/from tag are swapped inside force_rtp_proxy() before getting passed along to rtpproxy. The swap is critical as it provides consistent results when processing SDP offered in INVITE requests and Ok responses during re-invites within the same dialog. But the complimentary swapping is not being done for SDP answers contained in ACKs.

Closes patch 3019749
Credits go to John Riordan  


git-svn-id: https://opensips.svn.sourceforge.net/svnroot/opensips/trunk@7084 689a6050-402a-0410-94f2-e92a70836424